[explorer/frontend] feat: Add dual-chain variant shell
- Route shared API modules through the selected variant.
- Move existing NEM behavior into the NEM variant.
- Add minimal Symbol variant placeholders without page/API implementation.
- Move document metadata and page config behind the variant boundary.
