+++ explorer/frontend/setupTests.js
@@ -33,22 +33,35 @@ global.$t = key => `translated_${key}`;
 
 const originalEnv = { ...process.env };
 const envMock = {
-	NATIVE_MOSAIC_ID: 'nem.xem',
-	NATIVE_MOSAIC_TICKER: 'XEM',
-	NATIVE_MOSAIC_DIVISIBILITY: 6,
-	BLOCKCHAIN_UNWIND_LIMIT: 360,
-	REQUEST_TIMEOUT: 5000,
-	API_BASE_URL: 'https://explorer.backend',
-	SUPERNODE_API_URL: 'https://supernode.backend',
-	NODELIST_URL: 'https://node.list',
-	MARKET_DATA_URL: 'https://market.data',
-	HISTORICAL_PRICE_URL: 'https://historical.price'
+	NEXT_PUBLIC_PLATFORM: 'nem',
+	NEXT_PUBLIC_NATIVE_MOSAIC_ID: 'nem.xem',
+	NEXT_PUBLIC_NATIVE_MOSAIC_TICKER: 'XEM',
+	NEXT_PUBLIC_NATIVE_MOSAIC_DIVISIBILITY: 6,
+	NEXT_PUBLIC_BLOCKCHAIN_UNWIND_LIMIT: 360,
+	NEXT_PUBLIC_REQUEST_TIMEOUT: 5000,
+	NEXT_PUBLIC_API_BASE_URL: 'https://explorer.backend',
+	NEXT_PUBLIC_SUPERNODE_API_URL: 'https://supernode.backend',
+	NEXT_PUBLIC_NODELIST_URL: 'https://node.list',
+	NEXT_PUBLIC_MARKET_DATA_URL: 'https://market.data',
+	NEXT_PUBLIC_HISTORICAL_PRICE_URL: 'https://historical.price'
 };
 process.env = {
 	...originalEnv,
 	...envMock
 };
-window.appConfig = envMock;
+window.appConfig = {
+	PLATFORM: envMock.NEXT_PUBLIC_PLATFORM,
+	NATIVE_MOSAIC_ID: envMock.NEXT_PUBLIC_NATIVE_MOSAIC_ID,
+	NATIVE_MOSAIC_TICKER: envMock.NEXT_PUBLIC_NATIVE_MOSAIC_TICKER,
+	NATIVE_MOSAIC_DIVISIBILITY: envMock.NEXT_PUBLIC_NATIVE_MOSAIC_DIVISIBILITY,
+	BLOCKCHAIN_UNWIND_LIMIT: envMock.NEXT_PUBLIC_BLOCKCHAIN_UNWIND_LIMIT,
+	REQUEST_TIMEOUT: envMock.NEXT_PUBLIC_REQUEST_TIMEOUT,
+	API_BASE_URL: envMock.NEXT_PUBLIC_API_BASE_URL,
+	SUPERNODE_API_URL: envMock.NEXT_PUBLIC_SUPERNODE_API_URL,
+	NODELIST_URL: envMock.NEXT_PUBLIC_NODELIST_URL,
+	MARKET_DATA_URL: envMock.NEXT_PUBLIC_MARKET_DATA_URL,
+	HISTORICAL_PRICE_URL: envMock.NEXT_PUBLIC_HISTORICAL_PRICE_URL
+};
 
 Object.assign(global, { TextDecoder, TextEncoder });
