+++ explorer/frontend/variants/nem/config/pages/home.json
@@ -1,0 +1,4 @@
+{
+	"showSupernodeCount": true,
+	"additionalSections": []
+}
