+++ explorer/frontend/variants/nem/config/pages/index.js
@@ -1,0 +1,7 @@
+import home from './home.json';
+
+const pageConfig = {
+	home
+};
+
+export default pageConfig;
