+++ explorer/frontend/variants/nem/styles/variables.scss
@@ -1,0 +1,1 @@
+// NEM uses the shared default theme from styles/variables.scss.
