+++ explorer/frontend/variants/symbol/DocumentHead.jsx
@@ -1,0 +1,8 @@
+export const DocumentHead = () => (
+	<>
+		<meta charSet="utf-8" />
+		<meta name="format-detection" content="telephone=no" />
+		<meta name="theme-color" content="#1B0A29" />
+		<meta name="description" content="Symbol Block Explorer" />
+	</>
+);
