+++ explorer/frontend/variants/symbol/api/index.js
@@ -1,0 +1,29 @@
+const unsupportedSymbolApi = async () => {
+	throw new Error('Symbol API implementation belongs in the Symbol page/API implementation PR.');
+};
+
+export const fetchAccountPage = unsupportedSymbolApi;
+export const fetchAccountInfo = unsupportedSymbolApi;
+export const fetchAccountInfoByPublicKey = unsupportedSymbolApi;
+export const fetchBlockPage = unsupportedSymbolApi;
+export const fetchChainHeight = unsupportedSymbolApi;
+export const fetchBlockInfo = unsupportedSymbolApi;
+export const fetchMosaicPage = unsupportedSymbolApi;
+export const fetchMosaicInfo = unsupportedSymbolApi;
+export const fetchNamespacePage = unsupportedSymbolApi;
+export const fetchNamespaceInfo = unsupportedSymbolApi;
+export const fetchNodeList = unsupportedSymbolApi;
+export const search = unsupportedSymbolApi;
+export const fetchAccountStats = unsupportedSymbolApi;
+export const fetchTransactionChart = unsupportedSymbolApi;
+export const fetchTransactionStats = unsupportedSymbolApi;
+export const fetchBlockStats = unsupportedSymbolApi;
+export const fetchNodeStats = unsupportedSymbolApi;
+export const fetchMarketData = unsupportedSymbolApi;
+export const fetchPriceByDate = unsupportedSymbolApi;
+export const fetchTransactionPage = unsupportedSymbolApi;
+export const fetchTransactionInfo = unsupportedSymbolApi;
+export const resolveTransactionBlockSearch = unsupportedSymbolApi;
+export const resolveTransactionMosaicSearch = unsupportedSymbolApi;
+export const resolveTransactionRecipientSearch = unsupportedSymbolApi;
+export const resolveTransactionSignerSearch = unsupportedSymbolApi;
