+++ explorer/frontend/variants/symbol/config/index.js
@@ -1,0 +1,3 @@
+export { default as pageConfig } from './pages';
+
+export const config = {};
