+++ explorer/frontend/variants/symbol/config/pages/home.json
@@ -1,0 +1,4 @@
+{
+	"showSupernodeCount": false,
+	"additionalSections": []
+}
